+++ CHANGELOG.md
@@ -11,11 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GenerateDocumentationFile
 - Split files
 - Use EditorBrowsable(EditorBrowsableState.Never) instead of private or internal
+- Use Generic Comparer in BinarySearch, PriorityQueue
 
 ### Changed
 - Raise AC0008 on Method
 - Follow ac-library v1.3
-- Use Generic Comparer in BinarySearch, PriorityQueue
+- [Bug fix]Create all members in CreateOperatorCodeFixProvider
 
 ## [1.0.5] - 2020-12-18
 ### Added